okay save before change of plan
I was going to re-write the attributes of an abstract Comment, but i think that will make ALL button comments be changed to the same variable values....

so

NEXT
Im going to have button comments START with an undefined Comment object, which will then be given by minigameManager script. Well see how it goes.