Fixed text wrapping
Make sure in Text object the text wrapping is enabled. Overflow is optional, as it will 'overflow' text out of button bounds if must.
Secondly, enable auto size so that it will change the size if the text does not seem to fit.

The combination of the two will minimize the chances of overflowing up until about 3 lines. Should comment text exceed this however, there will be ugly overflow. To prevent that, I must either decrease the minimum size the auto scale option on the Text object cant take, or increase the button and text field size.